Add entries context and hook
